드림위치: 스팀 도전과제 연동 및 설정 완료, 보스전에서 힌트 대사 출력 해야할 것: 동굴 내부 배경 오브젝트 추가, 벽 점프 수정, 추격 및 동굴 브금 추가

+++ DreamWitch/Assets/Script/Title/SoundController.cs
@@ -13,7 +13,7 @@ public class SoundController : MonoBehaviour
     public AudioSource mBGM, mSE;
     public AudioClip[] mBGMArr, mSEArr;
     public AudioMixer mMixer;
-    public float BGMVolume, SEVolume;
+    public int BGMVolume, SEVolume;
 
     private void Awake()
     {
